==> 00-eda.ipynb <==
# Exploratory Data Analysis 

This notebook explores and transforms the multi-class data we will be using in the rest of the notebooks.

**Note:** you will not be able to run this notebook yourself unless you download the [raw data](https://kdd.ics.uci.edu/databases/20newsgroups/20newsgroups.html). This notebook specifically uses the '20_newsgroups' data set and assumes it is saved in a subdirectory of the working directory. 
Downloading this data set is *not* required for the rest of the notebooks in this series, so feel free to simply read through this notebook without executing the cells. 

The data set consists of 20,000 messages taken from 20 Usenet newsgroups. Each message is stored in its own file.

In the next few cells we inspect the contents of some of the files:

f = open('./20_newsgroups/alt.atheism/51121', 'r') 
lines = f.read()
print(lines)
f.close()

f = open('./20_newsgroups/comp.graphics/37921', 'r') 
print(lines)
f.close()

f = open('./20_newsgroups/comp.graphics/37930', 'r') 
lines = f.read()
print(lines)
f.close()

If we look at the data above we see some structure:

1. The file contains some information in headers.
2. The message itself starts after a double line break. 
3. There is information in the 'Keywords' and the 'Subject' lines of the header which may be informative about the classification of the message. 

We will check a few more files to see if they exhibit the same structure:

f = open('./20_newsgroups/comp.graphics/37916', 'r') 
lines = f.read()
print(lines)
f.close()

f = open('./20_newsgroups/misc.forsale/70337', 'r') 
lines = f.read()
print(lines)
f.close()

f = open('./20_newsgroups/misc.forsale/74797', 'r') 
lines = f.read()
print(lines)
f.close()

## Transforming the data

We want to store the data in `pandas` `DataFrame`, with each message making up a row of the DataFrame. 

We will do this by first creating a list containing this information, then we will make a DataFrame from the list. 

For each file we:
1. Extract the message by discarding everything before the first double line break,
2. Extract the `Subject:` line,
2. Extract the classification of the message from the filename. 

import os 

listOfFiles = list()
for (dirpath, dirnames, filenames) in os.walk('./20_newsgroups'):
    listOfFiles += [os.path.join(dirpath, file) for file in filenames]

data=[]
passed=[]
for file in listOfFiles:
    
    f = open(file, 'r') 
    try:
        lines = f.read()
        data.append([lines.split('\n\n', 1)[1], lines.split('\nSubject: ', 1)[1].split('\n', 1)[0], file.split('/')[2]])
    except:
        passed.append(file.split('/')[2:])
        pass
    f.close()

len(data)

We can look at a slice of the list like so:

data[0:10]

### Making a DataFrame

We transform the lit into a [Pandas](https://pandas.pydata.org) DataFrame as many machine learning models from python packages accept Pandas DataFrames as input. 

import pandas as pd

df = pd.DataFrame(data, columns = ['Message', 'Subject', 'Category'])

df.sample(10)

For simplicity, in the subsequent notebooks we want to treat the `Message` and the `Subject` as being part of one long piece of text. As such, we save time and computation by combining those DataFrame columns now:

df["Text"]=df["Message"]+df["Subject"]
df = df.drop(columns = ["Message", "Subject"])

df.sample(10)

### Splitting the data

We split the data, using 70% as a training set, and the remaining 30% as a testing set. We save these DataFrames  as parquet files.

train = df.sample(frac=0.7, random_state=504)
test = df.drop(train.index) #everything that isn't in the test set

train.shape

test.shape

train.to_parquet('data/training.parquet')

test.to_parquet('data/testing.parquet')
