==> 02-random-forest-model.ipynb <==
# Random Forest Classifier

In this notebook we train a [random forest classifier](https://en.wikipedia.org/wiki/Random_forest). Random forest classifiers are made up of sets of decision trees. As such, they are often deemed to be 'explainable' machine learning models.

### Loading in training data

import pandas as pd
import numpy as np

import os.path

training_data = pd.read_parquet(os.path.join("data", "training.parquet"))

training_data.sample(10)

### Feature Engineering

import cloudpickle as cp
feature_pipeline = cp.load(open('feature_pipeline.sav', 'rb'))

training_vecs = feature_pipeline.fit_transform(training_data["Text"])

### Training a Model

from sklearn.ensemble import RandomForestClassifier
from sklearn import model_selection

rfc = RandomForestClassifier(n_estimators=500, max_depth=5, random_state=404)

rfc.fit(training_vecs, training_data["Category"])

✅ Play around with the parameters of your random forest classifier. What happens to the performance as you increase `max_depth`?

### Evaluating model performance

rfc.score(training_vecs, training_data["Category"])

testing_data = pd.read_parquet(os.path.join("data", "testing.parquet"))
testing_vecs=feature_pipeline.transform(testing_data["Text"])
rfc.score(testing_vecs, testing_data["Category"])

from mlworkflows import plot

df, chart =plot.confusion_matrix(testing_data["Category"], rfc.predict(testing_vecs))

chart

from sklearn.metrics import classification_report
print(classification_report(testing_data["Category"], rfc.predict(testing_vecs)))

from mlworkflows import util

util.serialize_to(rfc, "model.sav")
